Quiz (4 panels, constant pairs) working, but note weird error in description.
Each quizBox has a table of 2 columns. Choosing matching item from left
and right column is success. Note that once a <div> appears to the
right of three rows of TD elements, those TD elements become
unclickable. Will move the scoreboard to be above the table but would
like a better workaround.